capstoneBeforeTests

+++ src/main/java/com/techelevator/ReservationDAO.java
@@ -4,6 +4,6 @@ import java.time.LocalDate;
 
 public interface ReservationDAO {
 	
-	public int makeReservation(String reserveName, CampSite selectedCampSite, LocalDate localDate1, LocalDate localdate2);
+	public Long makeReservation(String reserveName, CampSite selectedCampSite, LocalDate localDate1, LocalDate localdate2);
 
 }
